refactor: optimize display value logic and utilize useMemo for performance improvements
