+++ client/src/components/Chat/Menus/Endpoints/utils.ts
@@ -2,11 +2,12 @@ import React from 'react';
 import { Bot } from 'lucide-react';
 import { isAgentsEndpoint, isAssistantsEndpoint } from 'librechat-data-provider';
 import type {
+  TModelSpec,
   TAgentsMap,
   TAssistantsMap,
   TEndpointsConfig,
-  TModelSpec,
 } from 'librechat-data-provider';
+import type { useLocalize } from '~/hooks';
 import SpecIcon from '~/components/Chat/Menus/Endpoints/components/SpecIcon';
 import { Endpoint, SelectedValues } from '~/common';
 
@@ -155,3 +156,56 @@ export function getSelectedIcon({
 
   return null;
 }
+
+export const getDisplayValue = ({
+  localize,
+  mappedEndpoints,
+  selectedValues,
+  modelSpecs,
+}: {
+  localize: ReturnType<typeof useLocalize>;
+  selectedValues: SelectedValues;
+  mappedEndpoints: Endpoint[];
+  modelSpecs: TModelSpec[];
+}) => {
+  if (
+    selectedValues.modelSpec &&
+    !isAgentsEndpoint(selectedValues.endpoint) &&
+    !isAssistantsEndpoint(selectedValues.endpoint)
+  ) {
+    const spec = modelSpecs.find((s) => s.name === selectedValues.modelSpec);
+    return spec?.label || localize('com_ui_select_model');
+  }
+
+  if (selectedValues.model && selectedValues.endpoint) {
+    const endpoint = mappedEndpoints.find((e) => e.value === selectedValues.endpoint);
+    if (!endpoint) {
+      return localize('com_ui_select_model');
+    }
+
+    if (
+      isAgentsEndpoint(endpoint.value) &&
+      endpoint.agentNames &&
+      endpoint.agentNames[selectedValues.model]
+    ) {
+      return endpoint.agentNames[selectedValues.model];
+    }
+
+    if (
+      isAssistantsEndpoint(endpoint.value) &&
+      endpoint.assistantNames &&
+      endpoint.assistantNames[selectedValues.model]
+    ) {
+      return endpoint.assistantNames[selectedValues.model];
+    }
+
+    return selectedValues.model;
+  }
+
+  if (selectedValues.endpoint) {
+    const endpoint = mappedEndpoints.find((e) => e.value === selectedValues.endpoint);
+    return endpoint?.label || localize('com_ui_select_model');
+  }
+
+  return localize('com_ui_select_model');
+};
